+++ CHANGELOG.md
@@ -23,7 +23,7 @@ Resource data will be available with the resource object. This enhancement helps
 
   Refer example files of the updated resources for more details.
 
-#### features supported with current release
+#### Features supported with current release
 - Connection template
 - Enclosure
 - Enclosure group
@@ -35,6 +35,7 @@ Resource data will be available with the resource object. This enhancement helps
 - Managed SAN
 - SAS interconnect
 - SAS interconnect type
+- SAS logical interconnect
 - Server hardware type
 - Switch type
 - Uplink set